+++ cms-api/src/modules/block/block.route.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 
 import { asyncRouterHandler } from '../../errorHandling';
 import { injector } from '../../injector';
-import { validate } from '../../validation/validateMiddleware';
+import { validate } from '../../validation/validate.middleware';
 import { requiredParentId, updateFolderName, createFolder } from '../folder/folder.validation';
 import { requiredContentId, cutOrCopyContent, insertContent } from '../content/content.validation';
 import { BlockController } from './block.controller';
