+++ cms-api/src/modules/page/page.route.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { asyncRouterHandler } from '../../errorHandling';
 import { injector } from '../../injector';
-import { validate } from '../../validation/validateMiddleware';
+import { validate } from '../../validation/validate.middleware';
 import { requiredParentId } from '../folder/folder.validation';
 import { requiredContentId, cutOrCopyContent, insertContent } from '../content/content.validation';
