+++ cms-api/src/modules/user/user.route.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 
 import { asyncRouterHandler } from '../../errorHandling';
 import { injector } from '../../injector';
-import { validate } from '../../validation/validateMiddleware';
+import { validate } from '../../validation/validate.middleware';
 import { requiredId } from '../shared/base.validation';
 import { UserController } from './user.controller';
 import { createUser, updateUser } from './user.validation';
